\name{f_all_data_wide}
\alias{f_all_data_wide}

\title{Creates wide data for a given dataframe}

\description{This function produces wide data from long data and groups by the "symbol" column.}

\usage{
f_all_data_wide(t_input)}

\arguments{
  \item{t_input}{- Wide dataframe}
}

\value{Produces wide dataframe grouped by "symbol" from data passes through this function}
