\name{f_last_year}
\alias{f_last_year}

\title{Lag A Time Series.}

\description{Compute a lagged version of a time series, shifting the time base back by one year.}

\usage{
f_last_year(x)}

\arguments{
  \item{x}{- A vector or matrix or univariate or multivariate time series.}
}

\value{A time series object with the same class as x.}
